refactor: update rating data type and enhance ESLint configuration
- Changed rating data type from TINYINT to SMALLINT in migration and model files for better compatibility.
- Updated ESLint configuration to include additional plugins for React and accessibility.
- Added new ESLint dependencies in package.json and package-lock.json for improved code quality checks.

+++ backend/database/migrations/20260329113147-create-ratings.js
@@ -41,7 +41,7 @@ module.exports = {
         onDelete: 'SET NULL',
       },
       rating: {
-        type: Sequelize.TINYINT,
+        type: Sequelize.SMALLINT,
         allowNull: false,
         validate: {
           min: 1,
